Version con float

+++ main.cc
@@ -4,17 +4,20 @@
 
 class Exprexion {
 public:
-	virtual int evaluete(void) = 0;
+	virtual float evaluete(void) = 0;
 }; 
+
+
+
 class Literal :public Exprexion{
 	private:
-		int value;
+		float value;
 	public: 
-		Literal(int value){
+		Literal(float value){
 
 			this->value = value;
 		}
-		int evaluete(void){
+		float evaluete(void){
 
 			return value;
 		}
@@ -28,7 +31,7 @@ class Addition : public Exprexion{
 			this-> a = &a;
 			this-> b = &b;
 		}
-		int evaluete(void){
+		float evaluete(void){
 			return a->evaluete() + b->evaluete();
 		}	
 };
@@ -41,7 +44,7 @@ class Difference : public Exprexion{
 			this-> a = &a;
 			this-> b = &b;
 		}
-		int evaluete(void){
+		float evaluete(void){
 			return a->evaluete() - b->evaluete();
 		}	
 };
@@ -54,7 +57,7 @@ class Multiplication : public Exprexion{
 			this-> a = &a;
 			this-> b = &b;
 		}
-		int evaluete(void){
+		float evaluete(void){
 			return a->evaluete() * b->evaluete();
 		}	
 };
@@ -68,7 +71,7 @@ class Division : public Exprexion{
 			this-> a = &a;
 			this-> b = &b;
 		}
-		int evaluete(void){
+		float evaluete(void){
 			return a->evaluete() / b->evaluete();
 		}	
 };
@@ -83,13 +86,11 @@ int main(){
 	Literal num4(4);
 	Literal num5(5);
 	
-	
 	Multiplication mult(num3,num5);
 	Division divi(num4,num2);
 	Addition add(mult,divi);
 	Difference diff(add,num1);
 
 	printfExprexion(diff);
-	
 	return 0;
 };
